Deploy/requirements.txt
gradio==4.18.0
transformers==4.38.2
torch==2.2.2+cpu
torchvision==0.17.2+cpu
timm==0.9.12
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.4
scikit-learn==1.4.1.post1
boto3==1.34.91